Install node-json-db
As part of the reseed recipe implementation, I need to persist,
potentially among other things, the subcriptionId and webhookUrl. These
two are needed so we can know which webhook to call for which
"playground". In other words (subscriptionId, webhookUrl) uniquely
identifies the workspace to reseed.

I decided to use node-json-db b/c it's an embedded flat-file database
that has a simple API. The flat-file itself will be .gitignore'd, so
everyone can have a different db instance.